requirements_core.txt
pyyaml>=6.0
pydantic>=2.0.0
click>=8.0.0
requests>=2.28.0
neo4j>=5.0.0
openai>=1.0.0
python-dotenv>=1.0.0
